Move each viewstate to its own class
