+++ features/mpsearch/src/main/java/com/jpp/mpsearch/SearchPlaceHolderViewState.kt
@@ -1,0 +1,11 @@
+package com.jpp.mpsearch
+
+import android.view.View
+
+/**
+ * Represents the view state of the search placeholder view.
+ */
+data class SearchPlaceHolderViewState(
+    val visibility: Int = View.INVISIBLE,
+    val icon: Int = R.drawable.ic_app_icon_black
+)
