+++ features/mpsearch/src/main/java/com/jpp/mpsearch/SearchResultTypeIcon.kt
@@ -1,0 +1,11 @@
+package com.jpp.mpsearch
+
+import androidx.annotation.DrawableRes
+
+/**
+ * Represents the icon in the type of the [SearchResultItem].
+ */
+enum class SearchResultTypeIcon(@DrawableRes val iconRes: Int) {
+    Movie(R.drawable.ic_clapperboard),
+    Person(R.drawable.ic_person_black)
+}
